requirements.txt
py-cord
firebase-admin
python-dotenv~=1.0.1
requests~=2.32.0
pillow~=11.0.0